refactor : product 모듈분리
