+++ kitchenpos-product/src/main/java/kitchenpos/product/domain/Product.java
@@ -6,8 +6,8 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import kitchenpos.common.domain.Name;
-import kitchenpos.common.domain.Price;
+import kitchenpos.domain.Name;
+import kitchenpos.domain.Price;
 
 @Entity
 public class Product {
